requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-openai
langchain-google-genai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
beautifulsoup4
fastapi
uvicorn[standard]
langserve[all]
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools